# [WIP] Dynamic batch_size during inference