--- a/requirements/requirements.txt
+++ b/requirements/requirements.txt
@@ -1,4 +1,4 @@
 pyyaml==3.12
-mxnet-mkl==1.2.1
+mxnet-mkl>=1.2.1
 numpy>=1.14
 typing