Added option for --use-dynamic-batch-size, to allow Sockeye to support dynamic batch sizes during inference (on STDIN with json input only).

### Notes
This change makes the the "batch_size" argument dynamic, such that a user could specify a different batch_size per invocation. The default behavior of the pipeline is unchanged if you do not specify this option, and I have verified below that this does not have an impact on latency. I have also verified that there is no diff in the translation results whatsoever in the standard use-case.

This of course, may not be the case if we are to batch sentences together. If needed, we can do more extensive testing on that.

### Testing
Added new unit test cases, and everything passes.

For the dataset below:
```
[samples  202] [mean 18.30 tok] [+/- 12.61 tok] [max 123.00] [min 5.00]
```

#### Before Change
```
[samples  202] [mean 88.446089 tok/s] [+/- 22.044562 tok/s] [max 164.731684] [min 25.449224]
```

#### After Change
```
[samples  202] [mean 88.533140 tok/s] [+/- 22.097902 tok/s] [max 167.662940] [min 25.452589]
```

Note that this confirms that there should be no impact to latency.

#### Pull Request Checklist ##
- [ ] Changes are complete (if posting work-in-progress code, prefix your pull request title with '[WIP]'
until you can check this box.
- [ ] Unit tests pass (`pytest`)
- [x] Were system tests modified? If so did you run these at least 5 times to account for the variation across runs?
- [x] System tests pass (`pytest test/system`)
- [x] Passed code style checking (`./style-check.sh`)
- [x] You have considered writing a test
- [x] Updated major/minor version in `sockeye/__init__.py`. Major version bump if this is a backwards incompatible change.
- [x] Updated CHANGELOG.md


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.