@fhieber 2019-02-24
I am closing this in favor of #641. Thanks for getting this started!